fix path of undefined error
If `Router.current()` is "notFound", or some other route that is not handled by iron router, the original implementation will throw an error because `Router-current().route` is `undefined`. We can also use `Router.current()._handled` to determine whether it is defined, but I think it is supposed to be an interval variable, hence `Router.current().route`.